[#11] refactor: BaseEntity delete 기록 자동화 기능 수정
