+++ src/main/java/com/toyland/ToylandApplication.java
@@ -2,8 +2,9 @@ package com.toyland;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 public class ToylandApplication {
 
 	public static void main(String[] args) {
