+++ src/main/java/com/toyland/global/config/auditing/JpaAuditingConfig.java
@@ -1,16 +1,11 @@
 package com.toyland.global.config.auditing;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
 public class JpaAuditingConfig {
 
-    @Bean
-    public AuditorAware<Long> auditorAware() {
-        return new AuditorAwareImpl();
-    }
+
 }
